+++ src/solutions/solutionRunnerWorkerThread.js
